Disable fp16 due to NaN

+++ model/src/tensorflow/model.rs
@@ -21,7 +21,6 @@ use common::incrementing_map::IncrementingMap;
 use engine::game_state::GameState;
 use engine::engine::GameEngine;
 
-use super::half::Half;
 use super::constants::{ANALYSIS_REQUEST_BATCH_SIZE,ANALYSIS_REQUEST_THREADS,TRAIN_DATA_CHUNK_SIZE};
 use super::paths::Paths;
 use super::super::analytics::{self,ActionWithPolicy,GameStateAnalysis};
@@ -236,7 +235,7 @@ impl Predictor {
 
         let input_dim = self.input_dimensions;
         let input_dimensions = [batch_size as u64, input_dim[0], input_dim[1], input_dim[2]];
-        let flattened_inputs: Vec<Half> = game_state_inputs.iter().map(|v| v.iter()).flatten().map(|v| Half::from(*v)).collect();
+        let flattened_inputs: Vec<f32> = game_state_inputs.iter().map(|v| v.iter()).flatten().map(|v| f32::from(*v)).collect();
         let mut value_head_outputs: Vec<f32> = Vec::with_capacity(batch_size);
         let mut policy_head_outputs: Vec<f32> = Vec::with_capacity(batch_size);
 
@@ -250,8 +249,8 @@ impl Predictor {
 
         session.session.run(&mut output_step).unwrap();
 
-        let value_head_output: Tensor<Half> = output_step.fetch(value_head_fetch_token).unwrap();
-        let policy_head_output: Tensor<Half> = output_step.fetch(policy_head_fetch_token).unwrap();
+        let value_head_output: Tensor<f32> = output_step.fetch(value_head_fetch_token).unwrap();
+        let policy_head_output: Tensor<f32> = output_step.fetch(policy_head_fetch_token).unwrap();
 
         let policy_dimension = policy_head_output.dims()[1];
